apps/backend/requirements.txt
fastapi
uvicorn[standard]
pydantic
pydantic-settings
sqlalchemy
psycopg2-binary
redis
httpx
python-dotenv
chromadb
alembic
python-multipart
pymupdf
python-docx
sentence-transformers
numpy
openai
ollama
